Split CommonState type out of ConnectionCommon
This represents the state that state machine `State` implementers
can manipulate. As such, we want the lifetime of references to
`CommonState` to be outlived by the `ConnectionCommon`'s handshake
joiner and deframer. Now that it is split out of `ConnectionCommon`,
we can also replace the `ClientContext`/`ServerContext` indirection
with a parametrized `CommonState` type.